Mavenized the project; Renamed the S3 class to S3PutTask Added support for defining ContentType in S3PutTask Added empty unit test
