+++ README.md
@@ -28,7 +28,7 @@ Features:
 Added a very simple S3 task that enable files upload
 
 ```xml
-<taskdef name="s3" classpath="aws-ant-task.jar" classname="it.corley.ant.S3" />
+<taskdef name="s3" classpath="aws-ant-task.jar" classname="it.corley.ant.S3PutTask" />
 <s3 key="your-key" secret="your-secret" bucket="your-bucket-name" dest="path/to/file">
     <fileset dir="my/dir" includes="**/*.html" />
 </s3>
@@ -37,7 +37,7 @@ Added a very simple S3 task that enable files upload
 File upload using default [Ant fileset](http://ant.apache.org/manual/Types/fileset.html) strategy.
 
 ```xml
-<taskdef name="s3" classpath="aws-ant-task.jar" classname="it.corley.ant.S3" />
+<taskdef name="s3" classpath="aws-ant-task.jar" classname="it.corley.ant.S3PutTask" />
 <s3 key="your-key" secret="your-secret" bucket="your-bucket-name" dest="path/to/file">
   <fileset dir="${public.src}" casesensitive="yes">
     <patternset id="non.test.sources">
