+++ build.xml.dist
@@ -31,7 +31,7 @@
     
     <target name="use" description="Use the Task" depends="jar">
         <taskdef name="cloudfront" classpath="aws-ant-task-${version}.jar" classname="it.corley.ant.CloudFront" />
-        <taskdef name="s3" classpath="aws-ant-task-${version}.jar" classname="it.corley.ant.S3" />
+        <taskdef name="s3" classpath="aws-ant-task-${version}.jar" classname="it.corley.ant.S3PutTask" />
         <taskdef name="simpledb" classpath="aws-ant-task-${version}.jar" classname="it.corley.ant.SimpleDB" />
         
         <cloudfront key="your-key" secret="your-secret" distributionId="your-distribution">
