+++ src/test/it/corley/ant/S3PutTaskTest.java
@@ -1,0 +1,11 @@
+package it.corley.ant;
+
+/**
+ * Unit test class for the S3PutTask class
+ * Created at 11:26 10.05.12
+ *
+ * @author Tadeusz Kozak
+ * @see it.corley.ant.S3PutTask
+ */
+public class S3PutTaskTest {
+}
